chore: move cache files into folder
