+++ meteor/server/api/blueprints/context/adlibActions.ts
@@ -27,7 +27,7 @@ import { Rundown } from '../../../../lib/collections/Rundowns'
 import { RundownPlaylist, RundownPlaylistActivationId } from '../../../../lib/collections/RundownPlaylists'
 import { PieceInstance, wrapPieceToInstance } from '../../../../lib/collections/PieceInstances'
 import { PartInstanceId, PartInstance, PartInstances } from '../../../../lib/collections/PartInstances'
-import { CacheForRundownPlaylist } from '../../../DatabaseCaches'
+import { CacheForRundownPlaylist } from '../../../cache/DatabaseCaches'
 import { getResolvedPieces, setupPieceInstanceInfiniteProperties } from '../../playout/pieces'
 import { postProcessPieces, postProcessTimelineObjects } from '../postProcess'
 import { ShowStyleUserContext, UserContextInfo } from './context'
