+++ meteor/server/api/ingest/rundownInput.ts
@@ -120,8 +120,8 @@ import {
 } from '../playout/lib'
 import { PartInstances, PartInstance } from '../../../lib/collections/PartInstances'
 import { MethodContext } from '../../../lib/api/methods'
-import { CacheForRundownPlaylist, initCacheForRundownPlaylist } from '../../DatabaseCaches'
-import { prepareSaveIntoCache, savePreparedChangesIntoCache } from '../../DatabaseCache'
+import { CacheForRundownPlaylist, initCacheForRundownPlaylist } from '../../cache/DatabaseCaches'
+import { prepareSaveIntoCache, savePreparedChangesIntoCache } from '../../cache/DatabaseCache'
 import { reportRundownDataHasChanged } from '../asRunLog'
 import { Settings } from '../../../lib/Settings'
 import { AdLibAction } from '../../../lib/collections/AdLibActions'
