+++ meteor/server/api/peripheralDevice.ts
@@ -28,11 +28,15 @@ import { MethodContextAPI, MethodContext } from '../../lib/api/methods'
 import { triggerWriteAccess, triggerWriteAccessBecauseNoCheckNecessary } from '../security/lib/securityVerify'
 import { checkAccessAndGetPeripheralDevice } from './ingest/lib'
 import { PickerPOST } from './http'
-import { initCacheForNoRundownPlaylist, initCacheForRundownPlaylist, CacheForRundownPlaylist } from '../DatabaseCaches'
+import {
+	initCacheForNoRundownPlaylist,
+	initCacheForRundownPlaylist,
+	CacheForRundownPlaylist,
+} from '../cache/DatabaseCaches'
 import { RundownPlaylist } from '../../lib/collections/RundownPlaylists'
 import { getActiveRundownPlaylistsInStudio } from './playout/studio'
 import { StudioId } from '../../lib/collections/Studios'
-import { getValidActivationCache } from '../ActivationCache'
+import { getValidActivationCache } from '../cache/ActivationCache'
 import { UserActionsLog } from '../../lib/collections/UserActionsLog'
 import { PieceGroupMetadata } from '../../lib/rundown/pieces'
