+++ meteor/server/api/playout/__tests__/actions.test.ts
@@ -23,7 +23,7 @@ import {
 	wrapWithCacheForRundownPlaylistFromRundown,
 	wrapWithCacheForRundownPlaylist,
 	wrapWithCacheForRundownPlaylistFromStudio,
-} from '../../../DatabaseCaches'
+} from '../../../cache/DatabaseCaches'
 
 // const Timeline = mockupCollection(OrgTimeline)
