+++ meteor/server/api/playout/lookahead/__tests__/lookahead.test.ts
@@ -28,7 +28,7 @@ import { findLookaheadObjectsForPart } from '../findObjects'
 import { PartAndPieces, PartInstanceAndPieceInstances } from '../util'
 import { Random } from 'meteor/random'
 import { getPartId } from '../../../ingest/lib'
-import { wrapWithCacheForRundownPlaylist } from '../../../../DatabaseCaches'
+import { wrapWithCacheForRundownPlaylist } from '../../../../cache/DatabaseCaches'
 import { getLookeaheadObjects } from '..'
 import { SelectedPartInstancesTimelineInfo, SelectedPartInstanceTimelineInfo } from '../../timeline'
 import { testInFiber } from '../../../../../__mocks__/helpers/jest'
