+++ meteor/server/api/playout/lookahead/__tests__/util.test.ts
@@ -13,7 +13,7 @@ import { LookaheadMode, TSR } from '@sofie-automation/blueprints-integration'
 import { MappingsExt, Studios } from '../../../../../lib/collections/Studios'
 import { PartInstances, wrapPartToTemporaryInstance } from '../../../../../lib/collections/PartInstances'
 import _ from 'underscore'
-import { wrapWithCacheForRundownPlaylist } from '../../../../DatabaseCaches'
+import { wrapWithCacheForRundownPlaylist } from '../../../../cache/DatabaseCaches'
 import { testInFiber } from '../../../../../__mocks__/helpers/jest'
 import { getOrderedPartsAfterPlayhead } from '../util'
