+++ meteor/server/api/playout/timeline.ts
@@ -47,7 +47,7 @@ import { createPieceGroupFirstObject, getResolvedPiecesFromFullTimeline } from '
 import { PackageInfo } from '../../coreSystem'
 import { PartInstance, PartInstanceId } from '../../../lib/collections/PartInstances'
 import { PieceInstance } from '../../../lib/collections/PieceInstances'
-import { CacheForRundownPlaylist, CacheForStudioBase } from '../../DatabaseCaches'
+import { CacheForRundownPlaylist, CacheForStudioBase } from '../../cache/DatabaseCaches'
 import { PeripheralDeviceAPI } from '../../../lib/api/peripheralDevice'
 import { getExpectedLatency } from '../../../lib/collections/PeripheralDevices'
 import { processAndPrunePieceInstanceTimings, PieceInstanceWithTimings } from '../../../lib/rundown/infinites'
